MR-62: new styles, change from js to jsx
