+++ src/Components/Subscriptions/Table/table.module.css
@@ -1,11 +1,19 @@
 .classContainer {
   background-color: #0f232e;
   color: #f9a414;
-  font-size: 20px;
+  font-size: 15px;
   padding: 20px;
 }
 
 td {
   background-color: rgb(24, 19, 19);
   padding: 15px;
+}
+
+.button {
+  background-color: #f9a414;
+}
+
+.buttonDelete {
+  background-color: #e33613;
 }
